+++ airbyte-config/models/src/main/java/io/airbyte/config/Configs.java
@@ -33,6 +33,10 @@ public interface Configs {
 
   String getAirbyteVersion();
 
+  String getAirbyteApiUrl();
+
+  int getAirbyteApiPort();
+
   String getAirbyteVersionOrWarning();
 
   Path getConfigRoot();
@@ -46,6 +50,12 @@ public interface Configs {
   String getDatabasePassword();
 
   String getDatabaseUrl();
+
+  String getConfigDatabaseUser();
+
+  String getConfigDatabasePassword();
+
+  String getConfigDatabaseUrl();
 
   String getWebappUrl();
 
@@ -67,6 +77,9 @@ public interface Configs {
 
   String getKubeNamespace();
 
+  String getSubmitterNumThreads();
+
+  // Resources
   String getCpuRequest();
 
   String getCpuLimit();
@@ -75,7 +88,7 @@ public interface Configs {
 
   String getMemoryLimit();
 
-  // The following methods retrieve logging related information.
+  // Logging
   String getS3LogBucket();
 
   String getS3LogBucketRegion();
@@ -85,6 +98,10 @@ public interface Configs {
   String getAwsSecretAccessKey();
 
   String getS3MinioEndpoint();
+
+  String getGcpStorageBucket();
+
+  String getGoogleApplicationCredentials();
 
   enum TrackingStrategy {
     SEGMENT,