+++ airbyte-db/src/main/java/io/airbyte/db/Databases.java
@@ -25,11 +25,13 @@
 package io.airbyte.db;
 
 import io.airbyte.commons.lang.Exceptions;
+import io.airbyte.db.bigquery.BigQueryDatabase;
 import io.airbyte.db.jdbc.DefaultJdbcDatabase;
 import io.airbyte.db.jdbc.JdbcDatabase;
 import io.airbyte.db.jdbc.JdbcStreamingQueryConfiguration;
 import io.airbyte.db.jdbc.StreamingJdbcDatabase;
 import java.util.Optional;
+import java.util.function.Function;
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.jooq.SQLDialect;
 import org.slf4j.Logger;
@@ -43,7 +45,10 @@ public class Databases {
     return createDatabase(username, password, jdbcConnectionString, "org.postgresql.Driver", SQLDialect.POSTGRES);
   }
 
-  public static Database createPostgresDatabaseWithRetry(String username, String password, String jdbcConnectionString) {
+  public static Database createPostgresDatabaseWithRetry(String username,
+                                                         String password,
+                                                         String jdbcConnectionString,
+                                                         Function<Database, Boolean> isDbReady) {
     Database database = null;
 
     while (database == null) {
@@ -51,9 +56,10 @@ public class Databases {
 
       try {
         database = createPostgresDatabase(username, password, jdbcConnectionString);
-        Optional<String> uuid = ServerUuid.get(database);
-        if (uuid.isEmpty()) {
-          throw new Exception("Server UUID not available yet!");
+        if (!isDbReady.apply(database)) {
+          LOGGER.info("Database is not ready yet. Please wait a moment, it might still be initializing...");
+          database = null;
+          Exceptions.toRuntime(() -> Thread.sleep(5000));
         }
       } catch (Exception e) {
         // Ignore the exception because this likely means that the database server is still initializing.
@@ -71,6 +77,10 @@ public class Databases {
     return createJdbcDatabase(username, password, jdbcConnectionString, "com.amazon.redshift.jdbc.Driver");
   }
 
+  public static Database createMySqlDatabase(String username, String password, String jdbcConnectionString) {
+    return createDatabase(username, password, jdbcConnectionString, "com.mysql.cj.jdbc.Driver", SQLDialect.MYSQL);
+  }
+
   public static Database createSqlServerDatabase(String username, String password, String jdbcConnectionString) {
     return createDatabase(username, password, jdbcConnectionString, "com.microsoft.sqlserver.jdbc.SQLServerDriver", SQLDialect.DEFAULT);
   }
@@ -85,6 +95,18 @@ public class Databases {
                                         final String driverClassName,
                                         final SQLDialect dialect) {
     final BasicDataSource connectionPool = createBasicDataSource(username, password, jdbcConnectionString, driverClassName);
+
+    return new Database(connectionPool, dialect);
+  }
+
+  public static Database createDatabase(final String username,
+                                        final String password,
+                                        final String jdbcConnectionString,
+                                        final String driverClassName,
+                                        final SQLDialect dialect,
+                                        final String connectionProperties) {
+    final BasicDataSource connectionPool =
+        createBasicDataSource(username, password, jdbcConnectionString, driverClassName, Optional.ofNullable(connectionProperties));
 
     return new Database(connectionPool, dialect);
   }
@@ -143,6 +165,10 @@ public class Databases {
     connectionPool.setUrl(jdbcConnectionString);
     connectionProperties.ifPresent(connectionPool::setConnectionProperties);
     return connectionPool;
+  }
+
+  public static BigQueryDatabase createBigQueryDatabase(final String projectId, final String jsonCreds) {
+    return new BigQueryDatabase(projectId, jsonCreds);
   }
 
 }