+++ airbyte-server/src/main/java/io/airbyte/server/handlers/ConnectionsHandler.java
@@ -25,6 +25,7 @@
 package io.airbyte.server.handlers;
 
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableMap.Builder;
 import com.google.common.collect.Lists;
@@ -37,9 +38,9 @@ import io.airbyte.api.model.ConnectionSchedule;
 import io.airbyte.api.model.ConnectionStatus;
 import io.airbyte.api.model.ConnectionUpdate;
 import io.airbyte.api.model.ResourceRequirements;
-import io.airbyte.api.model.SyncMode;
 import io.airbyte.api.model.WorkspaceIdRequestBody;
 import io.airbyte.commons.enums.Enums;
+import io.airbyte.commons.json.Jsons;
 import io.airbyte.config.JobSyncConfig.NamespaceDefinitionType;
 import io.airbyte.config.Schedule;
 import io.airbyte.config.StandardDestinationDefinition;
@@ -49,12 +50,15 @@ import io.airbyte.config.helpers.ScheduleHelpers;
 import io.airbyte.config.persistence.ConfigNotFoundException;
 import io.airbyte.config.persistence.ConfigRepository;
 import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
+import io.airbyte.scheduler.persistence.WorkspaceHelper;
 import io.airbyte.server.converters.CatalogConverter;
 import io.airbyte.validation.json.JsonValidationException;
 import io.airbyte.workers.WorkerUtils;
 import java.io.IOException;
 import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
@@ -67,18 +71,50 @@ public class ConnectionsHandler {
 
   private final ConfigRepository configRepository;
   private final Supplier<UUID> uuidGenerator;
+  private final WorkspaceHelper workspaceHelper;
 
   @VisibleForTesting
-  ConnectionsHandler(final ConfigRepository configRepository, final Supplier<UUID> uuidGenerator) {
+  ConnectionsHandler(final ConfigRepository configRepository, final Supplier<UUID> uuidGenerator, final WorkspaceHelper workspaceHelper) {
     this.configRepository = configRepository;
     this.uuidGenerator = uuidGenerator;
+    this.workspaceHelper = workspaceHelper;
   }
 
-  public ConnectionsHandler(final ConfigRepository configRepository) {
-    this(configRepository, UUID::randomUUID);
+  public ConnectionsHandler(final ConfigRepository configRepository, final WorkspaceHelper workspaceHelper) {
+    this(configRepository, UUID::randomUUID, workspaceHelper);
+  }
+
+  private void validateWorkspace(UUID sourceId, UUID destinationId, Set<UUID> operationIds) {
+    final UUID sourceWorkspace = workspaceHelper.getWorkspaceForSourceId(sourceId);
+    final UUID destinationWorkspace = workspaceHelper.getWorkspaceForDestinationId(destinationId);
+
+    Preconditions.checkArgument(
+        sourceWorkspace.equals(destinationWorkspace),
+        String.format(
+            "Source and destination do not belong to the same workspace. Source id: %s, Source workspace id: %s, Destination id: %s, Destination workspace id: %s",
+            sourceId,
+            sourceWorkspace,
+            destinationId,
+            destinationWorkspace));
+
+    for (UUID operationId : operationIds) {
+      final UUID operationWorkspace = workspaceHelper.getWorkspaceForOperationId(operationId);
+      Preconditions.checkArgument(
+          sourceWorkspace.equals(operationWorkspace),
+          String.format(
+              "Operation and connection do not belong to the same workspace. Workspace id: %s, Operation id: %s, Operation workspace id: %s",
+              sourceWorkspace,
+              operationId,
+              operationWorkspace));
+    }
   }
 
   public ConnectionRead createConnection(ConnectionCreate connectionCreate) throws JsonValidationException, IOException, ConfigNotFoundException {
+    // Validate source and destination
+    configRepository.getSourceConnection(connectionCreate.getSourceId());
+    configRepository.getDestinationConnection(connectionCreate.getDestinationId());
+    validateWorkspace(connectionCreate.getSourceId(), connectionCreate.getDestinationId(), new HashSet<>(connectionCreate.getOperationIds()));
+
     final UUID connectionId = uuidGenerator.get();
 
     // persist sync
@@ -120,10 +156,6 @@ public class ConnectionsHandler {
       standardSync.withManual(true);
     }
 
-    // Validate source and destination
-    configRepository.getSourceConnection(connectionCreate.getSourceId());
-    configRepository.getDestinationConnection(connectionCreate.getDestinationId());
-
     configRepository.writeStandardSync(standardSync);
 
     trackNewConnection(standardSync);
@@ -133,8 +165,9 @@ public class ConnectionsHandler {
 
   private void trackNewConnection(final StandardSync standardSync) {
     try {
+      final UUID workspaceId = workspaceHelper.getWorkspaceForConnectionId(standardSync.getConnectionId());
       final Builder<String, Object> metadataBuilder = generateMetadata(standardSync);
-      TrackingClientSingleton.get().track("New Connection - Backend", metadataBuilder.build());
+      TrackingClientSingleton.get().track(workspaceId, "New Connection - Backend", metadataBuilder.build());
     } catch (Exception e) {
       LOGGER.error("failed while reporting usage.", e);
     }
@@ -167,7 +200,11 @@ public class ConnectionsHandler {
 
   public ConnectionRead updateConnection(ConnectionUpdate connectionUpdate) throws ConfigNotFoundException, IOException, JsonValidationException {
     // retrieve and update sync
-    final StandardSync persistedSync = configRepository.getStandardSync(connectionUpdate.getConnectionId())
+    final StandardSync persistedSync = configRepository.getStandardSync(connectionUpdate.getConnectionId());
+
+    validateWorkspace(persistedSync.getSourceId(), persistedSync.getDestinationId(), new HashSet<>(connectionUpdate.getOperationIds()));
+
+    final StandardSync newConnection = Jsons.clone(persistedSync)
         .withNamespaceDefinition(Enums.convertTo(connectionUpdate.getNamespaceDefinition(), NamespaceDefinitionType.class))
         .withNamespaceFormat(connectionUpdate.getNamespaceFormat())
         .withPrefix(connectionUpdate.getPrefix())
@@ -177,13 +214,13 @@ public class ConnectionsHandler {
 
     // update Resource Requirements
     if (connectionUpdate.getResourceRequirements() != null) {
-      persistedSync.withResourceRequirements(new io.airbyte.config.ResourceRequirements()
+      newConnection.withResourceRequirements(new io.airbyte.config.ResourceRequirements()
           .withCpuRequest(connectionUpdate.getResourceRequirements().getCpuRequest())
           .withCpuLimit(connectionUpdate.getResourceRequirements().getCpuLimit())
           .withMemoryRequest(connectionUpdate.getResourceRequirements().getMemoryRequest())
           .withMemoryLimit(connectionUpdate.getResourceRequirements().getMemoryLimit()));
     } else {
-      persistedSync.withResourceRequirements(WorkerUtils.DEFAULT_RESOURCE_REQUIREMENTS);
+      newConnection.withResourceRequirements(WorkerUtils.DEFAULT_RESOURCE_REQUIREMENTS);
     }
 
     // update sync schedule
@@ -191,12 +228,12 @@ public class ConnectionsHandler {
       final Schedule newSchedule = new Schedule()
           .withTimeUnit(toPersistenceTimeUnit(connectionUpdate.getSchedule().getTimeUnit()))
           .withUnits(connectionUpdate.getSchedule().getUnits());
-      persistedSync.withManual(false).withSchedule(newSchedule);
+      newConnection.withManual(false).withSchedule(newSchedule);
     } else {
-      persistedSync.withManual(true).withSchedule(null);
+      newConnection.withManual(true).withSchedule(null);
     }
 
-    configRepository.writeStandardSync(persistedSync);
+    configRepository.writeStandardSync(newConnection);
     return buildConnectionRead(connectionUpdate.getConnectionId());
   }
 
@@ -295,10 +332,6 @@ public class ConnectionsHandler {
 
   private StandardSync.Status toPersistenceStatus(ConnectionStatus apiStatus) {
     return Enums.convertTo(apiStatus, StandardSync.Status.class);
-  }
-
-  private SyncMode toApiSyncMode(io.airbyte.config.SyncMode persistenceStatus) {
-    return Enums.convertTo(persistenceStatus, SyncMode.class);
   }
 
   private ConnectionStatus toApiStatus(StandardSync.Status status) {