+++ airbyte-webapp/src/components/hooks/services/useConnectionHook.tsx
@@ -1,26 +1,27 @@
-import { useCallback, useEffect, useState } from "react";
+import { useCallback } from "react";
 import { useFetcher, useResource } from "rest-hooks";
 
-import config from "config";
-import { AnalyticsService } from "core/analytics/AnalyticsService";
-import ConnectionResource, {
+import FrequencyConfig from "config/FrequencyConfig.json";
+import {
   Connection,
+  ConnectionConfiguration,
+  ConnectionNamespaceDefinition,
+  connectionService,
+} from "core/domain/connection";
+
+import ConnectionResource, {
   ScheduleProperties,
 } from "core/resources/Connection";
 import { SyncSchema } from "core/domain/catalog";
 import { SourceDefinition } from "core/resources/SourceDefinition";
-import FrequencyConfig from "config/FrequencyConfig.json";
 import { Source } from "core/resources/Source";
 import { Routes } from "pages/routes";
 import useRouter from "../useRouterHook";
 import { Destination } from "core/resources/Destination";
-import useWorkspace from "./useWorkspaceHook";
-import {
-  ConnectionConfiguration,
-  ConnectionNamespaceDefinition,
-} from "core/domain/connection";
+import useWorkspace from "./useWorkspace";
 import { Operation } from "core/domain/connection/operation";
 import { equal } from "utils/objects";
+import { useAnalytics } from "components/hooks/useAnalytics";
 
 export type ValuesProps = {
   schedule: ScheduleProperties | null;
@@ -58,44 +59,25 @@ type UpdateStateConnection = {
   sourceName: string;
   prefix: string;
   connectionConfiguration: ConnectionConfiguration;
-  schedule: {
-    units: number;
-    timeUnit: string;
-  } | null;
+  schedule: ScheduleProperties | null;
 };
 
 export const useConnectionLoad = (
-  connectionId: string,
-  withRefresh?: boolean
-): { connection: Connection | null; isLoadingConnection: boolean } => {
-  const [connection, setConnection] = useState<null | Connection>(null);
-  const [isLoadingConnection, setIsLoadingConnection] = useState(false);
-
-  // TODO: change to useStatefulResource
-  const fetchConnection = useFetcher(ConnectionResource.detailShape(), false);
-  const baseConnection = useResource(ConnectionResource.detailShape(), {
+  connectionId: string
+): {
+  connection: Connection;
+  refreshConnectionCatalog: () => Promise<Connection>;
+} => {
+  const connection = useResource(ConnectionResource.detailShape(), {
     connectionId,
   });
 
-  useEffect(() => {
-    (async () => {
-      if (withRefresh) {
-        setIsLoadingConnection(true);
-        setConnection(
-          await fetchConnection({
-            connectionId,
-            withRefreshedCatalog: withRefresh,
-          })
-        );
-
-        setIsLoadingConnection(false);
-      }
-    })();
-  }, [connectionId, fetchConnection, withRefresh]);
+  const refreshConnectionCatalog = async () =>
+    await connectionService.getConnection(connectionId, true);
 
   return {
-    connection: withRefresh ? connection : baseConnection,
-    isLoadingConnection,
+    connection,
+    refreshConnectionCatalog,
   };
 };
 
@@ -108,6 +90,7 @@ const useConnection = (): {
 } => {
   const { push } = useRouter();
   const { finishOnboarding, workspace } = useWorkspace();
+  const analyticsService = useAnalytics();
 
   const createConnectionResource = useFetcher(ConnectionResource.createShape());
   const updateConnectionResource = useFetcher(ConnectionResource.updateShape());
@@ -136,7 +119,7 @@ const useConnection = (): {
         [
           [
             ConnectionResource.listShape(),
-            { workspaceId: config.ui.workspaceId },
+            { workspaceId: workspace.workspaceId },
             (
               newConnectionId: string,
               connectionsIds: { connections: string[] }
@@ -154,8 +137,7 @@ const useConnection = (): {
         equal(item.config, values.schedule)
       );
 
-      AnalyticsService.track("New Connection - Action", {
-        user_id: config.ui.workspaceId,
+      analyticsService.track("New Connection - Action", {
         action: "Set up connection",
         frequency: frequencyData?.text,
         connector_source_definition: source?.sourceName,