+++ airbyte-integrations/bases/source-acceptance-test/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.1.14
+Fix for NULL datetime in MySQL format (i.e. 0000-00-00): https://github.com/airbytehq/airbyte/pull/4465
+
+## 0.1.13
+Replace `validate_output_from_all_streams` with `empty_streams` param: https://github.com/airbytehq/airbyte/pull/4897
+
+## 0.1.12
+Improve error message when data mismatches schema: https://github.com/airbytehq/airbyte/pull/4753
+
+## 0.1.11
+Fix error in the naming of method `test_match_expected` for class `TestSpec`.
+
+## 0.1.10
+Add validation of input config.json against spec.json.
+
 ## 0.1.9
 Add configurable validation of schema for all records in BasicRead test: https://github.com/airbytehq/airbyte/pull/4345
 The validation is ON by default.